% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{LitSearch}
\alias{LitSearch}
\title{Search PubMed and export results as a .bib file}
\usage{
LitSearch(term)
}
\arguments{
\item{term}{Search term as defined by Entrez search usage. Function can't handle too broad terms which return more then 9999 results.}
}
\value{
A .bib file in the working directory with formatted search results.
}
\description{
Function utilizes Rentrez functional to search NCBI Pubmed and custom parser to export results into a .bib file ready to umport into Endnote or other soft.
}
\examples{
\dontrun{LitSearch("((animal migration[MeSH Terms] OR migrations[MeSH Terms]) AND(gene[All fields] OR gene[MeSH Terms] OR genetic[All fields] OR genetic[MeSH Terms] OR genetics[All fields] OR genetics[MeSH Terms]) AND (aves[ORGN] OR mammalia[ORGN]))")}
}
