% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{LocQual}
\alias{LocQual}
\title{A convinience function to check the usability of loci for population genetics analysis.}
\usage{
LocQual(
  genpop,
  file = T,
  OutputHW = T,
  OutputNull = T,
  OutputFis = T,
  OutputPerLoc = T,
  NullTreshold = 0.05
)
}
\arguments{
\item{genpop}{Genpop file or a genind object containing genotypes}

\item{file}{Set to TRUE if supllying genotypes in a file or to FALSE if in an already loaded genind object. Defaults to TRUE}

\item{OutputHW}{Set to TRUE to output results of Hardy-Weinberg analysis in a file. Defaults to TRUE}

\item{OutputNull}{Set to TRUE to output results of Null allele analysis in a file. Defaults to TRUE}

\item{OutputFis}{Set to TRUE to output results of Fis analysis in a file. Defaults to TRUE}

\item{OutputPerLoc}{Set to TRUE to output summary for each locus in a file. Defaults to TRUE}

\item{NullTreshold}{Integer. What null allele frequency count as significant. Defaults to 0.05}
}
\value{
A list containting raw results for three analysis and two tables with summary statistics.
Quicksummary shows the number and names of populations for each locus in which the deviation by all three checks occurs.
Qualitytest shows extended summary
HW, NuAl and Fis contain the results for individual checks.
Optionally parts of the analysis are outputed in the .txt files
}
\description{
Function checks the deviation from Hardy-Weinberg equilibrium, the significance of Fis deviation from zero and the presence of null-alleles for each supplied locus and population.
HW check and null-alleles are done using functions from packages pegas and PopGenREport. The Fis deviations is checked with Xi squared on the basis of Fis calculated with poppr.
If the locus shows deviation from HW, deviation of Fis from zero and null-alleles in half or more populations it's recommended to exclude it from population genetics analysis.
}
\examples{
TBD
}
