% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{MRatio}
\alias{MRatio}
\title{Calculate the Garza-Williamson index (M-ratio) for microsatellite loci}
\usage{
MRatio(genal, file = T, output = T, popnames = NULL)
}
\arguments{
\item{genal}{Genpop object or a genslex file to be read with poppr::read.genalex()}

\item{file}{Set to TRUE if supllying genotypes in a file or to FALSE if in an already loaded genind object. Defaults to TRUE}

\item{output}{Set to TRUE to output results into .txt file}

\item{popnames}{Names for the populations to use. If not supplied sequential numbers will be used.}
}
\value{
Named list with per locus and mean MRatio for for each pop
}
\description{
Function calculates Mratio for microsatellite loci suitable for the amplification fragment length analysis. The allelic range is calculated by automatically determining the motif on the basis of smallest difference between alleles.
}
\examples{
TBD
}
