% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{parse_summary_list}
\alias{parse_summary_list}
\title{Technical internal func}
\usage{
parse_summary_list(x)
}
\arguments{
\item{x}{tbd}
}
\value{
tbd
}
\description{
Technical internal func
}
\examples{
not needed
}
