% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIGGPopGen-package.R
\docType{package}
\name{VIGGPopGen-package}
\alias{VIGGPopGen}
\alias{VIGGPopGen-package}
\title{VIGGPopGen: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElPtiz/VIGGPopGen}
  \item Report bugs at \url{https://github.com/ElPtiz/VIGGPopGen/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
